Avoid postgresql auth failures in rb-site install
If postgresql is set up to use 'peer' or 'ident' authentication
methods, then unix username and database username need to match.
Thus we need to be running as 'reviewboard' when trying to log in
as 'reviewboard' database user. This takes some extra work to make
sure rb-site utility can create files and dirs in the appropriate
areas.